new session replace command and hotkey
Allows me to replace my current session with a new one. I would often
find myself creating a new project session and wanting to create the
old one.

Also removed the creation of the second window, since I don't always
use it, and manually creating it after creating the session is as quick as
switching to it.